feat(filmstaden): implement fetching of seat map
the dimensions returned from Filmstaden is sometimes a float, hence
the schema change.

Closes #687

+++ internal/graph/model/models_gen.go
@@ -90,8 +90,8 @@ type FilmstadenSeatCoordinates struct {
 }
 
 type FilmstadenSeatDimensions struct {
-	Width  int `json:"width"`
-	Height int `json:"height"`
+	Width  float64 `json:"width"`
+	Height float64 `json:"height"`
 }
 
 type FilmstadenSeatMap struct {
